T5AD-39 #comment fix: fix routing intent handler return issue
