+++ llm-backend/llm/dto.py
@@ -7,34 +7,23 @@ class ChatRequest(BaseModel):
     currentLocation: Optional[Dict[str, float]] = None
     currentTimestamp: Optional[int] = None
 
-
-class LegDTO(BaseModel):
-    type: str
-    durationInMinutes: int
-    busServiceNumber: Optional[str]
-    instruction: Optional[str]
-
-
-class RouteDTO(BaseModel):
-    durationInMinutes: int
-    legs: List[LegDTO]
-    summary: str
-    routeGeometry: Optional[str]
-
+class CoordinatesDTO(BaseModel):
+    latitude: float
+    longitude: float
 
 class DirectionsResponseDTO(BaseModel):
     type: str = "directions"
     startLocation: str
     endLocation: str
-    suggestedRoutes: List[RouteDTO]
-
+    startCoordinates: CoordinatesDTO
+    endCoordinates: CoordinatesDTO
+    suggestedRoutes: List[Any]
 
 class MessageResponseDTO(BaseModel):
     type: str = "message"
     message: str
     intent: Optional[str] = None
     slots: Optional[Dict[str, Any]] = None
-
 
 class ErrorResponseDTO(BaseModel):
     type: str = "error"
